refactoring/modularization
